requirements.txt
# AI & LLM
google-generativeai>=0.8.0
openai>=1.30.0
anthropic>=0.40.0
python-dotenv>=1.0.0

# Web Search & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# Resume & Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Database
sqlalchemy>=2.0.0

# CLI & UI
rich>=13.7.0
click>=8.1.0
questionary>=2.0.0

# Browser Automation
playwright>=1.40.0

# Async & Concurrency
aiohttp>=3.9.0
asyncio-contextmanager>=1.0.0

# Web (FastAPI frontend)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.0
sse-starlette>=2.0.0

# Django Admin
django>=5.0.0
httpx>=0.27.0

# Auth + crypto
bcrypt>=4.1.0
itsdangerous>=2.1.0

# Cloud DB
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Observability
sentry-sdk[fastapi]>=2.0.0

# HF deploy
huggingface_hub>=0.20.0

# Firebase Admin (verifies ID tokens server-side)
firebase-admin>=6.5.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.8.0
